+++ arches_lingo/src/arches_lingo/api.ts
@@ -731,6 +731,23 @@ export const importThesaurus = async (file: File, overwriteOption: string) => {
     }
 };
 
+export const fetchDereferencedResource = async (
+    resourceType: "concept" | "scheme",
+    resourceInstanceId: string,
+    format: string,
+): Promise<Blob> => {
+    const urlName =
+        resourceType === "concept"
+            ? "arches_lingo:concept"
+            : "arches_lingo:scheme";
+
+    const path = generateArchesURL(urlName, { id: resourceInstanceId });
+    const response = await fetch(`${path}?format=${format}`);
+
+    if (!response.ok) throw new Error(response.statusText);
+    return await response.blob();
+};
+
 export const exportThesaurus = async (
     resourceid: string,
     format: string,